passwordChange to another component
